backend/requirements.txt
Django==5.2.10
djangorestframework==3.16.1
psycopg2-binary==2.9.11
djangorestframework-simplejwt==5.5.1